+++ app/src/main/java/de/passwordvault/model/autofill/PasswordVaultAutofillService.java
@@ -1,0 +1,24 @@
+package de.passwordvault.model.autofill;
+
+import android.os.CancellationSignal;
+import android.service.autofill.AutofillService;
+import android.service.autofill.FillCallback;
+import android.service.autofill.FillRequest;
+import android.service.autofill.SaveCallback;
+import android.service.autofill.SaveRequest;
+import androidx.annotation.NonNull;
+
+
+public class PasswordVaultAutofillService extends AutofillService {
+
+    @Override
+    public void onFillRequest(@NonNull FillRequest fillRequest, @NonNull CancellationSignal cancellationSignal, @NonNull FillCallback fillCallback) {
+
+    }
+
+    @Override
+    public void onSaveRequest(@NonNull SaveRequest saveRequest, @NonNull SaveCallback saveCallback) {
+
+    }
+
+}